Converted tabs to spaces.
The project was a mixture of tabs and spaces.
This is controversial, but a choice needed to be made
and it's better to make a wholesale change at the start
of the project. Any changes going forward can how be
tracked as code changes only. Any historic diffs can
now use this revision as a baseline. No code changes
where made prior to this white space change.