Update DirectX version defines. Fix most build warnings.
Some version information was defined in Typedefs.h, which only
affected versions passed to the API. These #defines are also
used to select header features. We now use defaults, or specify
them explicitly in the project settings.

There is an issue with DirectInput, which can cause access
violations or affect system performance. This only seems to
be a problem with version 7 (which links against dinput.lib).
Version 8, which requires dinput8.lib seems to work fine on
Windows 8.x.