Remove extraneous header comments.
Older version control systems, like CVS, used keyword
substitution. This allowed commits to alter the source
and add content, usually in the form of comments. While
these have value, their value is reduced without the
corresponding source control history.

I considered leaving these in, but they don't add much
value, and their utility doesn't carry forward (new commits
are not included). Instead, I wrote a perl script to replace
them with the corresponding author names, weighted by the
number of commits. So the authors with the most commits to
a file appear first.

The result is that the original header and copyright is
included (even though the company is no longer around) and
the author names are retained. This allows us to keep proper
attribution, and to have the details in the history (all in
one commit), should anyone require them.