Remove Windows chrome in fullscreen
This project doesn't yet use true D3D fullsceen mode.
For the time being, we just remove the Windows chrome if
we're showing 'fullscreen' and we add it back when we
switch into windowed mode. If the new Windowed flag is
used to start, it will start without chrome, because the
window class (and thus default style) is set up prior to
parsing the command line. To switch modes, press the pause
key, then Alt+Enter. Also turned off the version text.